requirements.txt
numpy
pandas
pillow
opencv-python
matplotlib
scikit-learn
torch
torchvision
onnx
onnxruntime
tqdm
